refactor: convert BrowserPlugin to TS
